pyproject.toml
[tool.poetry]
name = "pyutils"
version = "0.1.0"
description = "Various utilities useful for Python projects"
repository = "https://github.com/a96tudor/pyutils"
authors = ["Tudor Avram <a96tudor@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
pymongo = "4.7.2"
certifi = "2024.2.2"
typing-extensions = "4.12.2"
hvac = "2.3.0"
pyyaml = "6.0.2"
sqlalchemy = "^2.0.41"
psycopg2-binary = "2.9.10"
rfc3339 = "^6.2"
pyjwt = "^2.10.1"
jinja2 = "^3.1.6"


[tool.poetry.dev-dependencies]
python = ">=3.10,<4.0"
pymongo = "4.7.2"
tox = "4.15.0"
pytest = "8.2.1"
hvac = "2.3.0"
pyyaml = "6.0.2"
types-PyYAML = "6.0.12.20250516"
sqlalchemy = "2.0.41"
psycopg2-binary = "2.9.10"


[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88
skip = ".git, .pyre, __pycache__, build, dist, venv, .tox"
